fix: change name to commentPostService
